[IMP] bantracking: create endpoint to get information from account.move
sequence_enfc is a field for manual writing of the sequence e-NFC
(electronic fiscal invoice number)
for external queries
Endpoint for account.move is created to get the required information
(default is set by MY_INVOICE_FIELDS variable). Context
'wrapped_search_service' is used to special format for different fields.

Also, it was added documentation and a test for the complete flow of a
sale order (creation and invoice).